Clustering information content of all partitions within a tree

Sums the clustering information content (Meila 2007) across each partition within a phylogenetic tree. This value will be greater than the total information content of the tree where a tree contains multiple partitions, as these partitions will contain mutual information.

ClusteringInfo(tree)

ClusteringInfoSplits(splits)

Arguments

tree
A tree of class phylo, a list of trees, or a multiPhylo object.

Value

Information content, in bits.

Details

Note that Meila (2007) and Vinh et al. (2010) deal with entropy, which denotes the bits necessary to denote the cluster to which each tip belongs, i.e. bits/tip (see Vinh et al. 2010). This function multiplies the entropy by the number of tips to produce a measure of the information content of the tree as a whole.

Functions

ClusteringInfoSplits: Calculate clustering information from splits instead of trees.

References

Meilă M (2007). “Comparing clusterings---an information based distance.” Journal of Multivariate Analysis, 98(5), 873--895. doi: 10.1016/j.jmva.2006.11.013 .