Solve linear assignment problem using LAPJV

Solve the Linear Assignment Problem with the algorithm of Jonker & Volgenant (1987).

LAPJV(x)

Arguments

x
Matrix to solve.

Value

A list with two entries: score, the score of the optimal matching; and matching, the columns matched to each row of the matrix in turn.

Details

The Linear Assignment Problem seeks to match each row of a matrix with a column, such that the cost of the matching is minimized.

The JV approach improves on the Hungarian algorithm, which is implemented in clue::solve_LSAP.

References

"A Shortest Augmenting Path Algorithm for Dense and Sparse Linear Assignment Problems," Computing 38, 325-340, 1987

R. Jonker and A. Volgenant, University of Amsterdam.