Approximate the NNI distance

An approximation of the Nearest Neighbour Interchange distance, based on Li et al. (1996).

NNIDist(tree1, tree2 = tree1)

Arguments

tree1, tree2
Single trees of class phylo to undergo comparison.

Value

NNIDist returns, for each pair of trees, a named vector containing three integers:

lower is a lower bound on the NNI distance, and corresponds to the RF distance between the trees.

tight_upper is an upper bound on the distance, based on calculated maximum diameters for trees with < 13 leaves. NA is returned if trees are too different to employ this approach.

loose_upper is a looser upper bound on the distance, using n log n + O(n).

Details

In brief, this approximation algorithm works by identifying edges in one tree that do not match edges in the second. Each of these edges must undergo at least one NNI operation in order to reconcile the trees. Edges that match in both trees need never undergo an NNI operation, and divide each tree into smaller regions. By 'cutting' matched edges into two, a tree can be divided into a number of regions that solely comprise unmatched edges.

These regions can be viewed as separate trees that need to be reconciled. One way to reconcile these trees is to conduct a series of NNI operations that reduce a tree to a pectinate (caterpillar) tree, then to conduct an analogue of the mergesort algorithm. This takes at most n log n + O(n) NNI operations, and provides a loose upper bound on the NNI score. The maximum number of moves for an n-leaf tree can be calculated exactly for small trees; this provides a tighter upper bound, but is unavailable for n > 12.

Leaves:	1	2	3	4	5	6	7	8	9	10	11	12	13

Diameter:	0	0	0	1	3	5	7	10	12	15	18	21	?

References

Li M, Tromp J, Zhang L (1996). “Some notes on the nearest neighbour interchange distance.” In Goos G, Hartmanis J, Leeuwen J, Cai J, Wong CK (eds.), Computing and Combinatorics, volume 1090, 343--351. Springer, Berlin, Heidelberg. ISBN 978-3-540-61332-9 978-3-540-68461-9, doi: 10.1007/3-540-61332-3_168 .

Examples

library('TreeTools')

NNIDist(BalancedTree(7), PectinateTree(7))

#>       lower tight_upper loose_upper 
#>           2           2          14 

NNIDist(BalancedTree(7), as.phylo(0:3, 7))

#>             [,1] [,2] [,3] [,4]
#> lower          4    4    4    4
#> tight_upper    7    7    7    7
#> loose_upper   18   18   18   18

NNIDist(as.phylo(0:3, 7), PectinateTree(7))

#>             [,1] [,2] [,3] [,4]
#> lower          4    4    4    4
#> tight_upper    7    7    7    7
#> loose_upper   18   18   18   18

NNIDist(list(BalancedTree(7), PectinateTree(7)),
        as.phylo(0:3, 7))

#> , , lower
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    4    4    4    4
#> [2,]    4    4    4    4
#> 
#> , , tight_upper
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    7    7    7    7
#> [2,]    7    7    7    7
#> 
#> , , loose_upper
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   18   18   18   18
#> [2,]   18   18   18   18
#> 

CompareAll(as.phylo(0:5, 8), NNIDist)

#> $lower
#>   1 2 3 4 5 6
#> 1   1 1 2 2 2
#> 2 1   1 2 2 2
#> 3 1 1   2 2 1
#> 4 2 2 2   1 1
#> 5 2 2 2 1   1
#> 6 2 2 1 1 1  
#> 
#> $tight_upper
#>   1 2 3 4 5 6
#> 1   1 1 3 3 3
#> 2 1   1 3 3 3
#> 3 1 1   3 3 1
#> 4 3 3 3   1 1
#> 5 3 3 3 1   1
#> 6 3 3 1 1 1  
#> 
#> $loose_upper
#>    1  2  3  4  5  6
#> 1     7  7 10 10 10
#> 2  7     7 10 10 10
#> 3  7  7    10 10  7
#> 4 10 10 10     7  7
#> 5 10 10 10  7     7
#> 6 10 10  7  7  7   
#>