Path distance

Calculate the path distance between trees.

PathDist(tree1, tree2 = NULL)

Arguments

tree1, tree2
Single trees of class phylo to undergo comparison.

Value

PathDist() returns a vector or distance matrix of distances between trees.

Details

This function is a wrapper for the function path.dist() in the phangorn package. It pre-processes trees to ensure that their internal representation does not cause the path.dist() function to crash R.

The path distance is also termed the cladistic difference or topological distance.

Use of the path distance is discouraged as it emphasizes shallow relationships at the expense of deeper (and arguably more fundamental) relationships (Farris, 1973).

References

Farris JS (1973). “On comparing the shapes of taxonomic trees.” Systematic Zoology, 22(1), 50--54. doi: 10.2307/2412378 .

See also

Examples

library('TreeTools')

PathDist(BalancedTree(7), PectinateTree(7))

#> [1] 4.690416

PathDist(BalancedTree(7), as.phylo(0:2, 7))

#> [1] 7.348469 6.782330 7.348469

PathDist(as.phylo(0:2, 7), PectinateTree(7))

#> [1] 7.745967 8.246211 8.124038

PathDist(list(bal = BalancedTree(7), pec = PectinateTree(7)),
        as.phylo(0:2, 7))

#> NULL

CompareAll(as.phylo(30:33, 8), PathDist)

#>          1        2        3        4
#> 1          3.872983 3.872983 5.567764
#> 2 3.872983          3.464102 6.928203
#> 3 3.872983 3.464102          6.928203
#> 4 5.567764 6.928203 6.928203