Quality of projected distances

ProjectionQuality() calculates the trustworthiness and continuity of projected distances (Venna & Kaski 2001; Kaski et al. 2003). Trustworthiness measures, on a scale from 0--1, the degree to which points that are nearby in a projection are truly close neighbours; continuity, the extent to which points that are truly nearby retain their close spatial proximity in a projection.

ProjectionQuality(original, projected, neighbours = 10L)

Arguments

original, projected
Square matrix or dist object containing original / projected pairwise distances.

neighbours
Number of nearest neighbours to use in calculation.

Value

ProjectionQuality() returns a named vector of length four, containing the entries: Trustworthiness, Continuity, TxC (the product of these values), and sqrtTxC (its square root).

References

Venna J, Kaski S (2001). “Neighborhood preservation in nonlinear projection methods: an experimental study.” In Dorffner G, Bischof H, Hornik K (eds.), Artificial Neural Networks — ICANN 2001, Lecture Notes in Computer Science, 485--491. doi: 10.1007/3-540-44668-0_68 .

Kaski S, Nikkilä J, Oja M, Venna J, Törönen P, Castrén E (2003). “Trustworthiness and metrics in visualizing similarity of gene expression.” BMC Bioinformatics, 4, 48. doi: 10.1186/1471-2105-4-48 .

See also

Author

Wrapper for functions from Charlotte Soneson's dreval, https://github.com/csoneson/dreval/blob/master/R/trustworthiness.R

Examples

library('TreeTools', quietly = TRUE, warn.conflict = FALSE)
trees <- as.phylo(0:10, nTip = 8)
distances <- ClusteringInfoDistance(trees)
projection <- cmdscale(distances)
ProjectionQuality(distances, dist(projection), 4)

#> Trustworthiness      Continuity             TxC         sqrtTxC 
#>       0.8080808       0.9040404       0.7305377       0.8547150