(Information-adjusted) Robinson-Foulds distance

Calculates the Robinson-Foulds distance, or the equivalent similarity measure, optionally annotating matched partitions and weighting partitions according to their phylogenetic information content.

RobinsonFouldsInfo(tree1, tree2, similarity = FALSE, normalize = FALSE,
  reportMatching = FALSE)

RobinsonFouldsInfoSplits(splits1, splits2, reportMatching = FALSE)

RobinsonFoulds(tree1, tree2, similarity = FALSE, normalize = FALSE,
  reportMatching = FALSE)

RobinsonFouldsSplits(splits1, splits2, reportMatching = FALSE)

Arguments

tree1
Trees of class phylo, with tips labelled identically, or lists of such trees to undergo pairwise comparison.

tree2
Trees of class phylo, with tips labelled identically, or lists of such trees to undergo pairwise comparison.

similarity
Logical specifying whether to report the result as a tree similarity, rather than a difference.

normalize
If a numeric value is provided, this will be used as a maximum value against which to rescale results. If TRUE, results will be rescaled against a maximum value calculated from the specified tree sizes and topology, as specified in the 'Normalization' section below. If FALSE, results will not be rescaled.

reportMatching
Logical specifying whether to return the clade matchings as an attribute of the score.

splits1
Logical matrices where each row corresponds to a terminal, either listed in the same order or bearing identical names (in any sequence), and each column corresponds to a bipartition split, such that each terminal is identified as a member of the ingroup (TRUE) or outgroup (FALSE) of the respective bipartition split.

splits2
Logical matrices where each row corresponds to a terminal, either listed in the same order or bearing identical names (in any sequence), and each column corresponds to a bipartition split, such that each terminal is identified as a member of the ingroup (TRUE) or outgroup (FALSE) of the respective bipartition split.

Details

RobinsonFoulds is an inefficient implementation of the Robinson-Foulds distance, included for use with VisualizeMatching. To generate the RF distance efficiently, use the function ape{treedist}.

Note that if reportMatching = TRUE, the pairScores attribute returns a logical matrix specifying whether each pair of partitions is identical.

RobinsonFouldsInfo calculates the tree similarity or distance by summing the phylogenetic information content of all partitions that are (or are not) identical in both trees. Consequently, partitions that are more likely to be identical by chance alone make a smaller contribution to overall tree distance, because their similarity is less remarkable.

Functions

RobinsonFouldsInfoSplits: Calculate information-adjusted Robinson-Foulds distance from splits instead of trees.

RobinsonFoulds: An inefficient implementation of the Robinson-Foulds distance, included for use with VisualizeMatching. To generate the RF distance efficiently, use the function ape{treedist}.

RobinsonFouldsSplits: Calculate Robinson-Foulds distance from splits instead of trees.

Normalization

RobinsonFoulds is normalized against the total number of partitions that are present.

RobinsonFouldsInfo is normalized against the sum of the phylogenetic information of all splits in both trees, treated independently.

References

Robinson DF, Foulds LR (1981). “Comparison of phylogenetic trees.” Mathematical Biosciences, 53(1-2), 131--147. ISSN 00255564, doi: 10.1016/0025-5564(81)90043-2 .

See also