Nye et al. (2006) tree comparison

Implements the tree comparison metric of Nye et al. (2006). In short, this finds the optimal matching that pairs each branch from one tree with a branch in the second, where matchings are scored according to the size of the largest bipartition that is consistent with both of them, normalized against the Jaccard index.

NyeTreeSimilarity(tree1, tree2, normalize = FALSE,
  reportMatching = FALSE)

NyeSplitSimilarity(splits1, splits2, normalize = TRUE,
  reportMatching = FALSE)

Arguments

tree1
Trees of class phylo, with tips labelled identically, or lists of such trees to undergo pairwise comparison.

tree2
Trees of class phylo, with tips labelled identically, or lists of such trees to undergo pairwise comparison.

normalize
If a numeric value is provided, this will be used as a maximum value against which to rescale results. If TRUE, results will be rescaled against a maximum value calculated from the specified tree sizes and topology, as specified in the 'Normalization' section below. If FALSE, results will not be rescaled.

reportMatching
Logical specifying whether to return the clade matchings as an attribute of the score.

splits1
Logical matrices where each row corresponds to a terminal, either listed in the same order or bearing identical names (in any sequence), and each column corresponds to a bipartition split, such that each terminal is identified as a member of the ingroup (TRUE) or outgroup (FALSE) of the respective bipartition split.

splits2
Logical matrices where each row corresponds to a terminal, either listed in the same order or bearing identical names (in any sequence), and each column corresponds to a bipartition split, such that each terminal is identified as a member of the ingroup (TRUE) or outgroup (FALSE) of the respective bipartition split.

Functions

NyeSplitSimilarity: Calculate tree similarity from splits instead of trees.

Normalization

If normalize = TRUE, then results will be rescaled from zero to one by dividing by the maximum possible value for trees of the given topologies, which is four less than the total number of nodes in both trees. You may wish to normalize instead against the maximum number of nodes present in a pair of trees with n terminals, by specifying normalize = <some number>.

References

Nye TMW, Liò P, Gilks WR (2006). “A novel algorithm and web-based tool for comparing two alternative phylogenetic trees.” Bioinformatics, 22(1), 117--119. doi: 10.1093/bioinformatics/bti720 .

See also